fix get_module
