+++ configs/cyclegan/cyclegan_lsgan_resnet_in_1x1_250k_summer2winter.py
@@ -29,13 +29,13 @@ model = dict(
             type='L1Loss',
             loss_weight=0.5,
             data_info=dict(
-                pred='identity_{domain_a}', target='real_{domain_a}'),
+                pred=f'identity_{domain_a}', target=f'real_{domain_a}'),
             reduction='mean'),
         dict(
             type='L1Loss',
             loss_weight=0.5,
             data_info=dict(
-                pred='identity_{domain_b}', target='real_{domain_b}'),
+                pred=f'identity_{domain_b}', target=f'real_{domain_b}'),
             reduction='mean')
     ])
 dataroot = './data/summer2winter_yosemite'
