+++ mmgen/models/translation_models/static_translation_gan.py
@@ -6,7 +6,7 @@ import torch.nn as nn
 from ..builder import MODELS, build_module
 from ..gans import BaseGAN
 from .base_translation_model import BaseTranslationModel
-
+from mmcv.parallel import MMDistributedDataParallel
 
 @MODELS.register_module()
 class StaticTranslationGAN(BaseTranslationModel, BaseGAN):
@@ -107,6 +107,24 @@ class StaticTranslationGAN(BaseTranslationModel, BaseGAN):
         # basic testing information
         self.batch_size = self.test_cfg.get('batch_size', 1)
 
+
+    def get_module(self, module):
+        """Get `nn.ModuleDict` to fit the `MMDistributedDataParallel`
+        interface.
+
+        Args:
+            module (MMDistributedDataParallel | nn.ModuleDict): The input
+                module that needs processing.
+
+        Returns:
+            nn.ModuleDict: The ModuleDict of multiple networks.
+        """
+        if isinstance(module, MMDistributedDataParallel):
+            return module.module
+
+        return module
+    
+    
     def _get_target_generator(self, domain):
         """get target generator."""
         assert self.is_domain_reachable(
@@ -114,7 +132,8 @@ class StaticTranslationGAN(BaseTranslationModel, BaseGAN):
         ), f'{domain} domain is not reachable, available domain list is\
             {self._reachable_domains}'
 
-        return self.generators[domain]
+        return self.get_module(self.generators)[domain]
+
 
     def _get_target_discriminator(self, domain):
         """get target discriminator."""
@@ -123,4 +142,4 @@ class StaticTranslationGAN(BaseTranslationModel, BaseGAN):
         ), f'{domain} domain is not reachable, available domain list is\
             {self._reachable_domains}'
 
-        return self.discriminators[domain]
+        return self.get_module(self.discriminators)[domain]
